# docs/tasky-api.swagger.yaml
openapi: 3.0.3
info:
  title: Tasky - OpenAPI 3.0
  version: 1.0.0
  description: |-
    Este es el API de la aplicación Tasky. 
    Algunos links útiles:
    - [The Tasky backend repository](https://github.com/wjfatuan/tasky-backend)
    - [The Tasky frontend repository](https://github.com/wjfatuan/tasky-frontend)
servers:
  - url: https://localhost:8000/api/v1
...
paths:
...
  /usuarios/{username}:
    get:
      tags:
        - usuario
      summary: Obtener un usuario por nombre de usuario
      description: ''
      parameters:
        - name: username
          in: path
          description: 'El nombre de usuario a buscar. El password se retorna vacío'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'          
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    patch:
      tags:
        - usuario
      summary: Update user
      description: Actualiza los datos del usuario. Solo el mismo usuario puede hacerlo.
      parameters:
        - name: username
          in: path
          description: name that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Actualizar un usuario existente en la aplicación
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        default:
          description: successful operation
    delete:
      tags:
        - usuario
      summary: Borrar usuario
      description: Solo lo puede hacer el mismo usuario
      parameters:
        - name: username
          in: path
          description: El usuario a borrar
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Nombre invalido
        '404':
          description: Usuario no encontrado
components:
  schemas:
...
    Usuario:
      allOf:
        - $ref: '#/components/schemas/Usuario'
        - type: object
          properties:
            uid:
              type: string
              format: uuid
              example: f81d4fae-7dec-11d0-a765-00a0c91e6bf5
...
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header